[Lee] (최적화) 주석 변경

+++ Back/src/main/java/com/application/safety/service/UserInfoService.java
@@ -24,12 +24,17 @@ public class UserInfoService {
     private final UserProfileRepository userProfileRepository;
     private final UserDataRepository userDataRepository;
 
-    // 근로자 등록 Create Service
+    /**
+     * 새로운 근로자 등록
+     * 등록에 필요한 데이터를 DTO로 받아서 저장
+     *
+     * @param dto 근로자의 정보를 담고 있는 요청 DTO
+     */
     @Transactional
     public void addUser(UserDTO.Request dto) {
         // UserProfile 새로운 객체 생성 후 값 삽입
         UserProfile userProfile = new UserProfile();
-        userProfile.setUserNo(dto.getUserNo());
+        userProfile.setUserNo((long) dto.getUserNo());
         userProfile.setUserName(dto.getUserName());
 
         UserProfile saveUserProfile = userProfileRepository.save(userProfile);
@@ -51,14 +56,19 @@ public class UserInfoService {
         userInfoRepository.save(userInfo);
     }
 
-    // 근로자 정보 전체 조회 Read Service
+    /**
+     * 모든 근로자의 정보를 전체 조회
+     * 각 근로자의 정보를 UserDTO.Response 객체로 변환하여 리스트로 반환
+     *
+     * @return 모든 근로자의 정보를 가지고 있는 UserDTO.Response 객체의 리스트
+     */
     @Transactional(readOnly = true)
     public List<UserDTO.Response> getUserInfoAll() {
         List<UserInfo> userInfoList = userInfoRepository.findAll();
 
         return userInfoList.stream().map(userInfo -> {
             UserDTO.Response dto = new UserDTO.Response();
-            dto.setUserNo(userInfo.getUserNo());
+            dto.setUserNo(Math.toIntExact(userInfo.getUserNo()));
             dto.setUserPosition(userInfo.getUserPosition());
             dto.setUserName(userInfo.getUserProfile().getUserName());
             dto.setUserAge(userInfo.getUserAge());
@@ -70,8 +80,14 @@ public class UserInfoService {
         }).collect(Collectors.toList());
     }
 
-    // 근로자 정보(상세) 조회
-    // 이름, 연령, 성별, 주민등록번호, 전화번호, 이메일, 주소, 직위, 입사일자, 은행명, 계좌번호
+
+    /**
+     * 근로자 정보 -상세 조회 (사원번호 userNo를 기준)
+     * 조회된 정보는 이름, 연령, 성별, 주민등록번호, 전화번호, 이메일, 주소, 직위, 입사 일자, 은행명, 계좌번호, 메모
+     *
+     * @param userNo 근로자의 고유 사원번호
+     * @return 근로자의 상세 정보를 담고 있는 UserInfoDTO 객체
+     */
     @Transactional(readOnly = true)
     public UserInfoDTO getUserInfo(int userNo) {
         UserInfo userInfo = getUserInfoEntity(userNo);
@@ -79,7 +95,7 @@ public class UserInfoService {
         UserInfoDTO userInfoDTO = new UserInfoDTO();
 
         // userProfile 조회 및 이름 설정
-        userInfoDTO.setUserNo(userInfo.getUserNo());
+        userInfoDTO.setUserNo(Math.toIntExact(userInfo.getUserNo()));
         userInfoDTO.setUserName(userProfile.getUserName());
 
         userInfoDTO.setUserAge(userInfo.getUserAge());
@@ -98,8 +114,15 @@ public class UserInfoService {
         return userInfoDTO;
     }
 
-    // 근로자 정보(상세) 수정
-    // 연령, 성별, 주민등록번호, 전화번호, 이메일, 주소, 직위, 입사일자, 은행명, 계좌번호
+
+    /**
+     * 근로자의 상세 정보 수정
+     * 수정할 데이터는 사원번호를 제외한 나머지 개인 정보
+     * 이름, 연령, 성별, 주민등록번호, 전화번호, 이메일, 주소, 직위, 입사 일자, 은행명, 계좌번호, 메모
+     *
+     * @param userInfoDTO 수정할 상세 정보 데이터를 담고 있는 DTO
+     * @return 수정된 사용자 정보를 담고 있는 UserInfo 객체
+     */
     @Transactional
     public UserInfo updateUserInfo(UserInfoDTO userInfoDTO) {
         UserInfo userInfo = getUserInfoEntity(userInfoDTO.getUserNo());
@@ -126,12 +149,16 @@ public class UserInfoService {
         return userInfo;
     }
 
-    // 사원 삭제
+    /**
+     * 근로자 삭제 (사원번호 userNo를 기준)
+     * 연관된 다른 엔티티(UserData, UserInfo)의 데이터도 삭제 처리
+     *
+     * @param userNo 삭제할 사원번호
+     */
     @Transactional
     public void deleteUser(int userNo) {
         UserProfile userProfile = userProfileRepository.findById(userNo).orElseThrow();
 
-        //cascade 보류, 기존방식으로 삭제
         userDataRepository.deleteByUserProfile(userProfile);
         userInfoRepository.deleteById(userNo);
         userProfileRepository.deleteById(userNo);
